+++ xmpp-client/src/main/java/org/xmpp/extension/jingle/JingleSession.java
@@ -3,5 +3,15 @@ package org.xmpp.extension.jingle;
 /**
  * @author Christian Schudt
  */
-public class JingleSession {
+public final class JingleSession {
+
+    private final String sessionId;
+
+    public JingleSession(String sessionId) {
+        this.sessionId = sessionId;
+    }
+
+    public String getSessionId() {
+        return sessionId;
+    }
 }